Rearrange handler functions
